Reorganized the frontend folder structure
