+++ client/src/assets/WarningIcon.tsx
