+++ client/src/components/PrivateRoute.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../contexts/AuthContext";
 
 const NotAllowed = () => {
   return (
