+++ client/src/contexts/AuthContext.tsx
