+++ client/src/pages/Login.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import { Form } from "../components/Form";
 import Input from "../components/Input";
-import { useAuth } from "../context/AuthContext";
+import { useAuth } from "../contexts/AuthContext";
 
 type FormData = {
   email: string;
